\name{DX.generator-package}
\alias{DX.generator-package}
\alias{DX.generator}
\docType{package}
\title{
  Generates Random Unifrom Numbers using the DX algorthim
}
\description{
  This package is designed to provide a faster alternative to the default RNGs within R. This random number generator is .......
}
\details{
  
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Lih-Yuan Deng

Maintainer: Jon McCurdy <jonmccurdy@ymail.com>
}
\references{
  Will link a wedbsite explaining the DX-generator
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
    set.seed(1234)
    dx_runif(10)
}
