% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_runif.R
\name{dx_runif}
\alias{dx_runif}
\title{DX_runif}
\usage{
dx_runif(n, k = 47, s = 1, t = 1)
}
\arguments{
\item{n}{number of random numbers wanted}

\item{k}{DX explanation}

\item{s}{DX explanation}

\item{t}{DX explanation}
}
\value{

}
\description{
DX_runif
}
\examples{
dx_runif(10)

}
