% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rand_test}
\alias{rand_test}
\title{Gets the Random Numbers}
\usage{
rand_test(nTry, seed, k)
}
\arguments{
\item{nTry}{number of observations}

\item{seed}{seed to configure RNG}

\item{k}{DX parameter}
}
\value{
nTry random numbers
}
\description{
Gets the Random Numbers
}
